chore: package, compilation and workflow general changes
## `package.json` changes:

frontend -> @animegakuen/cosplay-backend
backend -> @animegakuen/cosplay-frontend

flagged `@animegakuen/cosplay-backend` as a private package, also set
it's version to `0.0.0`.

## nix package changes:

Done the appopriate re-naming for the packages, `pnpmWorkspaces` and on
compilation.

## github actions:

Set workflows to only run when pushed to the `main` branch and corrected
the workflow paths with a recursive match (`**`) inside each package's
workflow.